move `compute_r_powers_for_verify_kzg_proof_batch` into cryptography module
